+++ plugins/voxelman/client/plugin.d
@@ -7,7 +7,6 @@ module voxelman.client.plugin;
 
 import core.thread : thread_joinAll;
 import core.time;
-import std.datetime : StopWatch;
 import std.experimental.logger;
 
 import dlib.math.vector;
@@ -28,6 +27,7 @@ import voxelman.gui.plugin;
 import voxelman.net.plugin;
 import voxelman.command.plugin;
 import voxelman.block.plugin;
+import voxelman.world.plugin;
 
 import voxelman.net.events;
 import voxelman.core.packets;
@@ -43,11 +43,9 @@ import voxelman.storage.coordinates;
 import voxelman.storage.utils;
 import voxelman.storage.worldaccess;
 import voxelman.utils.math;
-import voxelman.utils.trace : traceRay;
 import voxelman.utils.textformatter;
 
 import voxelman.client.appstatistics;
-import voxelman.client.chunkman;
 import voxelman.client.console;
 
 //version = manualGC;
@@ -60,12 +58,6 @@ shared static this()
 	auto c = new ClientPlugin;
 	pluginRegistry.regClientPlugin(c);
 	pluginRegistry.regClientMain(&c.run);
-}
-
-struct ThisClientLoggedInEvent {
-	ClientId thisClientId;
-	Profiler profiler;
-	bool continuePropagation = true;
 }
 
 auto formatDuration(Duration dur)
@@ -86,18 +78,12 @@ private:
 	GraphicsPlugin graphics;
 	GuiPlugin guiPlugin;
 	CommandPluginClient commandPlugin;
-
-	// Resource managers
-	KeyBindingManager keyBindingMan;
-	ConfigManager config;
+	ClientWorld clientWorld;
+	NetClientPlugin connection;
 
 public:
 	AppStatistics stats;
 	Console console;
-
-	// Game stuff
-	ChunkMan chunkMan;
-	WorldAccess worldAccess;
 
 	// Debug
 	Profiler profiler;
@@ -105,66 +91,36 @@ public:
 
 	// Client data
 	bool isRunning = false;
-	bool isDisconnecting = false;
-	bool isSpawned = false;
 	bool mouseLocked;
 
-	ConfigOption serverIpOpt;
-	ConfigOption serverPortOpt;
 	ConfigOption runDespikerOpt;
-	ConfigOption numWorkersOpt;
-	ConfigOption nicknameOpt;
 
 	// Graphics stuff
 	bool isCullingEnabled = true;
-	bool doUpdateObserverPosition = true;
-	vec3 updatedCameraPos;
 	bool isConsoleShown = false;
-
-	// Client id stuff
-	ClientId thisClientId;
-	string[ClientId] clientNames;
-
-	// Send position interval
-	double sendPositionTimer = 0;
-	enum sendPositionInterval = 0.1;
-	ChunkWorldPos prevChunkPos;
-
-	NetClientPlugin connection;
 
 	// IPlugin stuff
 	mixin IdAndSemverFrom!(voxelman.client.plugininfo);
 
 	override void registerResources(IResourceManagerRegistry resmanRegistry)
 	{
-		config = resmanRegistry.getResourceManager!ConfigManager;
-		keyBindingMan = resmanRegistry.getResourceManager!KeyBindingManager;
-
-		serverIpOpt = config.registerOption!string("ip", "127.0.0.1");
-		serverPortOpt = config.registerOption!ushort("port", 1234);
+		ConfigManager config = resmanRegistry.getResourceManager!ConfigManager;
 		runDespikerOpt = config.registerOption!bool("run_despiker", false);
-		numWorkersOpt = config.registerOption!uint("num_workers", 4);
-		nicknameOpt = config.registerOption!string("name", "Player");
 
+		KeyBindingManager keyBindingMan = resmanRegistry.getResourceManager!KeyBindingManager;
 		keyBindingMan.registerKeyBinding(new KeyBinding(KeyCode.KEY_Q, "key.lockMouse", null, &onLockMouse));
-		keyBindingMan.registerKeyBinding(new KeyBinding(KeyCode.KEY_RIGHT_BRACKET, "key.incViewRadius", null, &onIncViewRadius));
-		keyBindingMan.registerKeyBinding(new KeyBinding(KeyCode.KEY_LEFT_BRACKET, "key.decViewRadius", null, &onDecViewRadius));
 		keyBindingMan.registerKeyBinding(new KeyBinding(KeyCode.KEY_C, "key.toggleCulling", null, &onToggleCulling));
-		keyBindingMan.registerKeyBinding(new KeyBinding(KeyCode.KEY_U, "key.togglePosUpdate", null, &onTogglePositionUpdate));
 		keyBindingMan.registerKeyBinding(new KeyBinding(KeyCode.KEY_GRAVE_ACCENT, "key.toggle_console", null, &onConsoleToggleKey));
 	}
 
 	override void preInit()
 	{
-		worldAccess.onChunkModifiedHandlers ~= &chunkMan.onChunkChanged;
 		console.init();
 	}
 
 	override void init(IPluginManager pluginman)
 	{
-		BlockPlugin blockPlugin = pluginman.getPlugin!BlockPlugin;
-		chunkMan.init(numWorkersOpt.get!uint, blockPlugin.getBlocks());
-
+		clientWorld = pluginman.getPlugin!ClientWorld;
 		evDispatcher = pluginman.getPlugin!EventDispatcherPlugin;
 		evDispatcher.profiler = profiler;
 
@@ -176,39 +132,15 @@ public:
 		evDispatcher.subscribeToEvent(&drawScene);
 		evDispatcher.subscribeToEvent(&drawOverlay);
 		evDispatcher.subscribeToEvent(&onClosePressedEvent);
-		evDispatcher.subscribeToEvent(&onGameStopEvent);
-		evDispatcher.subscribeToEvent(&handleThisClientConnected);
-		evDispatcher.subscribeToEvent(&handleThisClientDisconnected);
 
 		commandPlugin = pluginman.getPlugin!CommandPluginClient;
-		commandPlugin.registerCommand("connect", &connectCommand);
 		console.messageWindow.messageHandler = &onConsoleCommand;
 
 		connection = pluginman.getPlugin!NetClientPlugin;
-
-		connection.printPacketMap();
-
-		connection.registerPacketHandler!PacketMapPacket(&handlePacketMapPacket);
-		connection.registerPacketHandler!SessionInfoPacket(&handleSessionInfoPacket);
-		connection.registerPacketHandler!ClientLoggedInPacket(&handleUserLoggedInPacket);
-		connection.registerPacketHandler!ClientLoggedOutPacket(&handleUserLoggedOutPacket);
-		connection.registerPacketHandler!ClientPositionPacket(&handleClientPositionPacket);
-		connection.registerPacketHandler!ChunkDataPacket(&handleChunkDataPacket);
-		connection.registerPacketHandler!MultiblockChangePacket(&handleMultiblockChangePacket);
-		connection.registerPacketHandler!SpawnPacket(&handleSpawnPacket);
 	}
 
 	override void postInit()
 	{
-		updatedCameraPos = graphics.camera.position;
-		chunkMan.updateObserverPosition(graphics.camera.position);
-		ConnectionSettings settings = {null, 1, 2, 0, 0};
-
-		connection.start(settings);
-		static if (ENABLE_RLE_PACKET_COMPRESSION)
-			enet_host_compress_with_range_coder(connection.host);
-		connect(serverIpOpt.get!string, serverPortOpt.get!ushort);
-
 		if (runDespikerOpt.get!bool)
 			toggleProfiler();
 	}
@@ -233,7 +165,7 @@ public:
 			igTextf("Pos: X %.2f, Y %.2f, Z %.2f", pos.x, pos.y, pos.z);
 		}
 
-		ChunkWorldPos chunkPos = chunkMan.observerPosition;
+		ChunkWorldPos chunkPos = clientWorld.chunkMan.observerPosition;
 		auto regionPos = RegionWorldPos(chunkPos);
 		auto localChunkPosition = ChunkRegionPos(chunkPos);
 		igTextf("C: %s R: %s L: %s", chunkPos, regionPos, localChunkPosition);
@@ -242,18 +174,8 @@ public:
 		vec2 heading = graphics.camera.heading;
 		igTextf("Heading: %.2f %.2f Target: X %.2f, Y %.2f, Z %.2f",
 			heading.x, heading.y, target.x, target.y, target.z);
-		igTextf("Chunks to remove: %s", chunkMan.removeQueue.length);
-		igTextf("Chunks to mesh: %s", chunkMan.chunkMeshMan.numMeshChunkTasks);
-
-		igSeparator();
-		if (igButton("Profiler"))
-			toggleProfiler();
-		igSameLine();
-		if (igButton("Stop server"))
-			connection.send(CommandPacket("sv_stop"));
-		igSameLine();
-		if (igButton("Connect"))
-			connect(serverIpOpt.get!string, serverPortOpt.get!ushort);
+		igTextf("Chunks to remove: %s", clientWorld.chunkMan.removeQueue.length);
+		igTextf("Chunks to mesh: %s", clientWorld.chunkMan.chunkMeshMan.numMeshChunkTasks);
 		igEnd();
 	}
 
@@ -267,7 +189,6 @@ public:
 			profiler = new Profiler(storage);
 		}
 		profilerSender = new DespikerSender([profiler]);
-		worldAccess = WorldAccess(&chunkMan.chunkStorage.getChunk, () => 0);
 	}
 
 	void load(string[] args)
@@ -291,6 +212,7 @@ public:
 		version(manualGC) GC.disable;
 
 		load(args);
+		evDispatcher.postEvent(GameStartEvent());
 
 		TickDuration lastTime = Clock.currAppTick;
 		TickDuration newTime = TickDuration.from!"seconds"(0);
@@ -340,35 +262,7 @@ public:
 		}
 		profilerSender.reset();
 
-		isDisconnecting = connection.isConnected;
-		connection.disconnect();
-
-		infof("disconnecting");
-		while (isDisconnecting)
-		{
-			connection.update();
-		}
-		infof("stop");
-
 		evDispatcher.postEvent(GameStopEvent());
-	}
-
-	void connect(string ip, ushort port)
-	{
-		console.lineBuffer.putfln("Connecting to %s:%s", ip, port);
-		if (connection.isConnecting)
-			connection.disconnect();
-		connection.connect(ip, port);
-	}
-
-	void connectCommand(CommandParams params)
-	{
-		short port = serverPortOpt.get!ushort;
-		string serverIp = serverIpOpt.get!string;
-		getopt(params.args,
-			"ip", &serverIp,
-			"port", &port);
-		connect(serverIp, port);
 	}
 
 	void toggleProfiler()
@@ -385,63 +279,19 @@ public:
 		}
 	}
 
-	void onGameStopEvent(ref GameStopEvent gameStopEvent)
-	{
-		chunkMan.stop();
-		thread_joinAll();
-	}
-
 	void onPreUpdateEvent(ref PreUpdateEvent event)
 	{
-		if (doUpdateObserverPosition)
-		{
-			updatedCameraPos = graphics.camera.position;
-		}
-		chunkMan.updateObserverPosition(updatedCameraPos);
-		connection.update();
-		chunkMan.update();
 	}
 
 	void onPostUpdateEvent(ref PostUpdateEvent event)
 	{
-		updateStats();
+		stats.fps = guiPlugin.fpsHelper.fps;
+		stats.totalLoadedChunks = clientWorld.chunkMan.totalLoadedChunks;
+
 		printDebug();
 		stats.resetCounters();
 		if (isConsoleShown)
 			console.draw();
-		if (doUpdateObserverPosition)
-			sendPosition(event.deltaTime);
-		connection.flush();
-	}
-
-	void sendPosition(double dt)
-	{
-		ChunkWorldPos chunkPos = BlockWorldPos(graphics.camera.position);
-
-		if (isSpawned)
-		{
-			sendPositionTimer += dt;
-			if (sendPositionTimer > sendPositionInterval ||
-				chunkPos != prevChunkPos)
-			{
-				connection.send(ClientPositionPacket(
-					graphics.camera.position,
-					graphics.camera.heading));
-
-				if (sendPositionTimer < sendPositionInterval)
-					sendPositionTimer = 0;
-				else
-					sendPositionTimer -= sendPositionInterval;
-			}
-		}
-
-		prevChunkPos = chunkPos;
-	}
-
-	void updateStats()
-	{
-		stats.fps = guiPlugin.fpsHelper.fps;
-		stats.totalLoadedChunks = chunkMan.totalLoadedChunks;
 	}
 
 	void onConsoleCommand(string command)
@@ -467,38 +317,6 @@ public:
 		isConsoleShown = !isConsoleShown;
 	}
 
-	void incViewRadius()
-	{
-		setViewRadius(getViewRadius() + 1);
-	}
-
-	void decViewRadius()
-	{
-		setViewRadius(getViewRadius() - 1);
-	}
-
-	int getViewRadius()
-	{
-		return chunkMan.viewRadius;
-	}
-
-	void setViewRadius(int newViewRadius)
-	{
-		auto oldViewRadius = chunkMan.viewRadius;
-		chunkMan.viewRadius = clamp(newViewRadius,
-			MIN_VIEW_RADIUS, MAX_VIEW_RADIUS);
-
-		if (oldViewRadius != chunkMan.viewRadius)
-		{
-			connection.send(ViewRadiusPacket(chunkMan.viewRadius));
-		}
-	}
-
-	void sendMessage(string msg)
-	{
-		connection.send(MessagePacket(0, msg));
-	}
-
 	void onClosePressedEvent(ref ClosePressedEvent event)
 	{
 		isRunning = false;
@@ -511,24 +329,9 @@ public:
 			guiPlugin.window.mousePosition = cast(ivec2)(guiPlugin.window.size) / 2;
 	}
 
-	void onIncViewRadius(string)
-	{
-		incViewRadius();
-	}
-
-	void onDecViewRadius(string)
-	{
-		decViewRadius();
-	}
-
 	void onToggleCulling(string)
 	{
 		isCullingEnabled = !isCullingEnabled;
-	}
-
-	void onTogglePositionUpdate(string)
-	{
-		doUpdateObserverPosition = !doUpdateObserverPosition;
 	}
 
 	void drawScene(ref Render1Event event)
@@ -548,9 +351,9 @@ public:
 		frustum.fromMVP(vp);
 
 		Matrix4f modelMatrix;
-		foreach(ChunkWorldPos cwp; chunkMan.chunkMeshMan.visibleChunks.items)
+		foreach(ChunkWorldPos cwp; clientWorld.chunkMan.chunkMeshMan.visibleChunks.items)
 		{
-			Chunk* c = chunkMan.getChunk(cwp);
+			Chunk* c = clientWorld.chunkMan.getChunk(cwp);
 			assert(c);
 			++stats.chunksVisible;
 
@@ -585,98 +388,5 @@ public:
 		//event.renderer.setColor(Color(0,0,0,1));
 		//event.renderer.fillRect(Rect(guiPlugin.window.size.x/2-7, guiPlugin.window.size.y/2-1, 14, 2));
 		//event.renderer.fillRect(Rect(guiPlugin.window.size.x/2-1, guiPlugin.window.size.y/2-7, 2, 14));
-	}
-
-	void handleThisClientConnected(ref ThisClientConnectedEvent event)
-	{
-		infof("Connection to %s:%s established", serverIpOpt.get!string, serverPortOpt.get!ushort);
-	}
-
-	void handleThisClientDisconnected(ref ThisClientDisconnectedEvent event)
-	{
-		infof("disconnected with data %s", event.data);
-
-		isDisconnecting = false;
-		isSpawned = false;
-	}
-
-	void handlePacketMapPacket(ubyte[] packetData, ClientId clientId)
-	{
-		auto packetMap = unpackPacket!PacketMapPacket(packetData);
-
-		connection.setPacketMap(packetMap.packetNames);
-		connection.printPacketMap();
-
-		connection.send(ViewRadiusPacket(chunkMan.viewRadius));
-		connection.send(LoginPacket(nicknameOpt.get!string));
-	}
-
-	void handleSessionInfoPacket(ubyte[] packetData, ClientId clientId)
-	{
-		auto loginInfo = unpackPacket!SessionInfoPacket(packetData);
-
-		clientNames = loginInfo.clientNames;
-		thisClientId = loginInfo.yourId;
-		evDispatcher.postEvent(ThisClientLoggedInEvent(thisClientId));
-	}
-
-	void handleUserLoggedInPacket(ubyte[] packetData, ClientId clientId)
-	{
-		auto newUser = unpackPacket!ClientLoggedInPacket(packetData);
-		clientNames[newUser.clientId] = newUser.clientName;
-		infof("%s has connected", newUser.clientName);
-		evDispatcher.postEvent(ClientLoggedInEvent(clientId));
-	}
-
-	void handleUserLoggedOutPacket(ubyte[] packetData, ClientId clientId)
-	{
-		auto packet = unpackPacket!ClientLoggedOutPacket(packetData);
-		infof("%s has disconnected", clientName(packet.clientId));
-		evDispatcher.postEvent(ClientLoggedOutEvent(clientId));
-		clientNames.remove(packet.clientId);
-	}
-
-	void handleClientPositionPacket(ubyte[] packetData, ClientId peer)
-	{
-		import voxelman.utils.math : nansToZero;
-
-		auto packet = unpackPacket!ClientPositionPacket(packetData);
-		tracef("Received ClientPositionPacket(%s, %s)",
-			packet.pos, packet.heading);
-
-		nansToZero(packet.pos);
-		graphics.camera.position = packet.pos;
-
-		nansToZero(packet.heading);
-		graphics.camera.setHeading(packet.heading);
-	}
-
-	void handleSpawnPacket(ubyte[] packetData, ClientId peer)
-	{
-		auto packet = unpackPacket!SpawnPacket(packetData);
-		isSpawned = true;
-	}
-
-	void handleChunkDataPacket(ubyte[] packetData, ClientId peer)
-	{
-		auto packet = unpackPacket!ChunkDataPacket(packetData);
-		//tracef("Received %s ChunkDataPacket(%s,%s)", packetData.length,
-		//	packet.chunkPos, packet.blockData.blocks.length);
-		chunkMan.onChunkLoaded(ChunkWorldPos(packet.chunkPos), packet.blockData);
-	}
-
-	void handleMultiblockChangePacket(ubyte[] packetData, ClientId peer)
-	{
-		auto packet = unpackPacket!MultiblockChangePacket(packetData);
-		Chunk* chunk = chunkMan.chunkStorage.getChunk(ChunkWorldPos(packet.chunkPos));
-		// We can receive data for chunk that is already deleted.
-		if (chunk is null || chunk.isMarkedForDeletion)
-			return;
-		chunkMan.onChunkChanged(chunk, packet.blockChanges);
-	}
-
-	string clientName(ClientId clientId)
-	{
-		return clientId in clientNames ? clientNames[clientId] : format("? %s", clientId);
 	}
 }