+++ plugins/voxelman/clientdb/clientinfo.d
@@ -20,4 +20,5 @@ struct ClientInfo
 	Volume visibleVolume;
 
 	bool isLoggedIn;
+	bool isSpawned;
 }