+++ plugins/voxelman/net/plugin.d
@@ -6,6 +6,7 @@ Authors: Andrey Penechko.
 
 module voxelman.net.plugin;
 
+import std.experimental.logger;
 import derelict.enet.enet;
 
 import pluginlib;
@@ -19,6 +20,7 @@ import voxelman.core.packets;
 import voxelman.net.packets;
 
 import voxelman.eventdispatcher.plugin;
+import voxelman.command.plugin;
 import voxelman.config.configmanager;
 
 shared static this()
@@ -49,6 +51,11 @@ mixin template NetCommon()
 
 final class NetClientPlugin : IPlugin
 {
+	CommandPluginClient commandPlugin;
+	ConfigOption serverIpOpt;
+	ConfigOption serverPortOpt;
+	bool isDisconnecting = false;
+
 	mixin NetCommon;
 
 	BaseClient connection;
@@ -58,13 +65,113 @@ final class NetClientPlugin : IPlugin
 		connection = new class BaseClient{};
 	}
 
+	override void registerResources(IResourceManagerRegistry resmanRegistry)
+	{
+		ConfigManager config = resmanRegistry.getResourceManager!ConfigManager;
+
+		serverIpOpt = config.registerOption!string("ip", "127.0.0.1");
+		serverPortOpt = config.registerOption!ushort("port", 1234);
+	}
+
+	override void init(IPluginManager pluginman)
+	{
+		evDispatcher = pluginman.getPlugin!EventDispatcherPlugin;
+		evDispatcher.subscribeToEvent(&handleGameStartEvent);
+		evDispatcher.subscribeToEvent(&onPreUpdateEvent);
+		evDispatcher.subscribeToEvent(&onPostUpdateEvent);
+		evDispatcher.subscribeToEvent(&onGameStopEvent);
+		evDispatcher.subscribeToEvent(&handleThisClientConnected);
+		evDispatcher.subscribeToEvent(&handleThisClientDisconnected);
+
+		commandPlugin = pluginman.getPlugin!CommandPluginClient;
+		commandPlugin.registerCommand("connect", &connectCommand);
+
+		connection.registerPacketHandler!PacketMapPacket(&handlePacketMapPacket);
+	}
+
+	override void postInit()
+	{
+	}
+
+	void handleGameStartEvent(ref GameStartEvent event)
+	{
+		ConnectionSettings settings = {null, 1, 2, 0, 0};
+
+		connection.start(settings);
+		static if (ENABLE_RLE_PACKET_COMPRESSION)
+			enet_host_compress_with_range_coder(connection.host);
+		connect(serverIpOpt.get!string, serverPortOpt.get!ushort);
+	}
+
+	void connect(string ip, ushort port)
+	{
+		infof("Connecting to %s:%s", ip, port);
+		if (connection.isConnecting)
+			connection.disconnect();
+		connection.connect(ip, port);
+	}
+
+	void connectCommand(CommandParams params)
+	{
+		short port = serverPortOpt.get!ushort;
+		string serverIp = serverIpOpt.get!string;
+		getopt(params.args,
+			"ip", &serverIp,
+			"port", &port);
+		connect(serverIp, port);
+	}
+
+	void onPreUpdateEvent(ref PreUpdateEvent event)
+	{
+		connection.update();
+	}
+
+	void onPostUpdateEvent(ref PostUpdateEvent event)
+	{
+		connection.flush();
+	}
+
 	void onConnect(ref ENetEvent event) {
-		evDispatcher.postEvent(ThisClientConnectedEvent());
 	}
 
 	void onDisconnect(ref ENetEvent event) {
 		event.peer.data = null;
 		evDispatcher.postEvent(ThisClientDisconnectedEvent(event.data));
+	}
+
+	void handleThisClientConnected(ref ThisClientConnectedEvent event)
+	{
+		infof("Connection to %s:%s established", serverIpOpt.get!string, serverPortOpt.get!ushort);
+	}
+
+	void handleThisClientDisconnected(ref ThisClientDisconnectedEvent event)
+	{
+		infof("disconnected with data %s", event.data);
+		isDisconnecting = false;
+	}
+
+	void handlePacketMapPacket(ubyte[] packetData, ClientId clientId)
+	{
+		auto packetMap = unpackPacket!PacketMapPacket(packetData);
+		connection.setPacketMap(packetMap.packetNames);
+		connection.printPacketMap();
+
+		evDispatcher.postEvent(ThisClientConnectedEvent());
+		evDispatcher.postEvent(SendClientSettingsEvent());
+		connection.send(GameStartPacket());
+	}
+
+	void onGameStopEvent(ref GameStopEvent gameStopEvent)
+	{
+		isDisconnecting = connection.isConnected;
+		connection.disconnect();
+
+		infof("disconnecting");
+		while (isDisconnecting)
+		{
+			connection.update();
+		}
+		infof("stop");
 	}
 }